+++ src/Debug.php
@@ -98,6 +98,11 @@ class Debug
         if ($truncated && mb_strlen($response, 'utf8') > 1000) {
             $response = mb_substr($response, 0, 1000, 'utf8').'...';
         }
+
+        if (ctype_print($response) === false) {
+            $response = 'RAW DATA';
+        }
+
         if ($cliDebug) {
             echo $res.$response."\n\n";
         }